src/main/java/uk/gov/companieshouse/authcodenotification/exception/GlobalExceptionHandler.java
package uk.gov.companieshouse.authcodenotification.exception;

import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;
import org.springframework.web.context.request.WebRequest;
import uk.gov.companieshouse.authcodenotification.utils.ApiLogger;

import static uk.gov.companieshouse.authcodenotification.utils.Constants.ERIC_REQUEST_ID_KEY;

@ControllerAdvice
public class GlobalExceptionHandler {

    @ExceptionHandler(Exception.class)
    public ResponseEntity<Object> handleException(Exception ex, WebRequest webRequest) {

        var context = webRequest.getHeader(ERIC_REQUEST_ID_KEY);

        ApiLogger.errorContext(context, "Unhandled exception", ex);
        return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
    }
}
